+++ src/commands/q-mongo.ts
@@ -64,7 +64,7 @@ export async function qMongoBranch(
     return
   }
 
-  const adapter = AdapterFactory.createMongoDBAdapter(config.connection as ConnectionOptions)
+  const adapter = AdapterFactory.createMongoDBAdapter(config)
   await adapter.connect()
   try {
     const start = performance.now()